dalsze prace

+++ src/main/java/pl/wsikora/successbudget/v3/category/infrastructure/CategoryCommandImpl.java
@@ -4,43 +4,44 @@ import org.springframework.stereotype.Service;
 import pl.wsikora.successbudget.v3.category.application.CategoryAttributes;
 import pl.wsikora.successbudget.v3.category.application.CategoryCommand;
 import pl.wsikora.successbudget.v3.category.domain.Category;
-import pl.wsikora.successbudget.v3.common.type.*;
-import pl.wsikora.successbudget.v3.user.domain.User;
+import pl.wsikora.successbudget.v3.common.type.Description;
+import pl.wsikora.successbudget.v3.common.type.Title;
+import pl.wsikora.successbudget.v3.common.type.TransactionType;
+import pl.wsikora.successbudget.v3.common.type.Username;
 
 
 @Service
 class CategoryCommandImpl implements CategoryCommand {
 
-    private final UserCrudRepository userCrudRepository;
     private final CategoryRepository categoryRepository;
 
-    private CategoryCommandImpl(UserCrudRepository userCrudRepository,
-                                CategoryRepository categoryCrudRepository) {
+    private CategoryCommandImpl(CategoryRepository categoryCrudRepository) {
 
-        this.userCrudRepository = userCrudRepository;
         this.categoryRepository = categoryCrudRepository;
     }
 
     @Override
     public void save(CategoryAttributes categoryAttributes, Username username) {
 
-        User user = userCrudRepository.findById(username)
-            .orElseThrow(() -> new IllegalArgumentException("No user found for username: " + username));
+        Category category = new Category();
 
-        Category category = Category.builder()
-            .categoryId(new CategoryId(categoryAttributes.getCategoryId()))
-            .title(new Title(categoryAttributes.getTitle()))
-            .description(new Description(categoryAttributes.getDescription()))
-            .assignedTransactionType(TransactionType.valueOf(categoryAttributes.getAssignedTransactionType()))
-            .ownedByUser(user)
-            .build();
+        category.setCategoryId(categoryAttributes.getCategoryId());
+        category.setTitle(new Title(categoryAttributes.getTitle()));
+        category.setDescription(new Description(categoryAttributes.getDescription()));
+        category.setOwner(username);
+
+        TransactionType transactionType = TransactionType.values()[categoryAttributes.getAssignedTransactionType()];
+        category.setAssignedTransactionType(transactionType);
 
         categoryRepository.save(category);
     }
 
     @Override
-    public void delete(CategoryId categoryId) {
+    public void delete(Long categoryId, Username username) {
 
-        categoryRepository.deleteById(categoryId);
+        if (categoryRepository.existsByCategoryIdAndUsername(categoryId, username)) {
+
+            categoryRepository.deleteById(categoryId);
+        }
     }
 }
